# 8342975: C2: Micro-optimize PhaseIdealLoop::Dominators()