@shipilev 2024-10-24
On various tests on x86_64, this gives me +1% faster runs in `-Xcomp` scenarios. Not very visible with "normal" amount of C2 compilations.

```
## HelloWorld, -Xcomp -XX:-TieredCompilation 

# Before
  Time (mean ± σ):     617.7 ms ±   2.5 ms    [User: 584.6 ms, System: 31.5 ms]
  Range (min … max):   614.2 ms … 624.5 ms    20 runs

# After
  Time (mean ± σ):     611.3 ms ±   1.9 ms    [User: 578.0 ms, System: 31.8 ms]
  Range (min … max):   608.0 ms … 614.4 ms    20 runs
 

## JavacBenchApp 50, -XX:-TieredCompilation

# Before
  Time (mean ± σ):      1.733 s ±  0.011 s    [User: 3.074 s, System: 0.139 s]
  Range (min … max):    1.719 s …  1.753 s    20 runs

# After
  Time (mean ± σ):      1.727 s ±  0.011 s    [User: 3.023 s, System: 0.144 s]
  Range (min … max):    1.704 s …  1.751 s    20 runs


## JavacBenchApp 50, -Xcomp -XX:-TieredCompilation

# Before
  Time (mean ± σ):     15.223 s ±  0.061 s    [User: 15.048 s, System: 0.239 s]
  Range (min … max):   15.152 s … 15.330 s    10 runs

# After
  Time (mean ± σ):     15.080 s ±  0.021 s    [User: 14.896 s, System: 0.239 s]
  Range (min … max):   15.048 s … 15.115 s    10 runs
```